feat: updated index
